+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Sending WalletConnect transaction with 0 energy if its payload is too large
 
+### Removed
+- Shielding
+
 ## [1.5.1] - 2024-03-18
 
 ### Fixed