+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.6.0] - 2024-06-26
 
 ### Fixed
 - Sending WalletConnect transaction with 0 energy if its payload is too large
@@ -119,7 +119,8 @@ but to unshield the funds CryptoX Concordium wallet must be used
 
 - Last release without changelog.
 
-[Unreleased]: https://github.com/Concordium/concordium-reference-wallet-android/compare/1.5.1...HEAD
+[Unreleased]: https://github.com/Concordium/concordium-reference-wallet-android/compare/1.6.0...HEAD
+[1.6.0]: https://github.com/Concordium/concordium-reference-wallet-android/compare/1.5.1...1.6.0
 [1.5.1]: https://github.com/Concordium/concordium-reference-wallet-android/compare/1.5.0...1.5.1
 [1.5.0]: https://github.com/Concordium/concordium-reference-wallet-android/compare/1.3.0...1.5.0
 [1.3.0]: https://github.com/Concordium/concordium-reference-wallet-android/compare/1.2.1...1.3.0